+++ ios/RNInstabug/ArgsRegistry.m
@@ -21,7 +21,8 @@
     [all addEntriesFromDictionary:ArgsRegistry.nonFatalExceptionLevel];
     [all addEntriesFromDictionary:ArgsRegistry.placeholders];
     [all addEntriesFromDictionary:ArgsRegistry.launchType];
-
+    [all addEntriesFromDictionary:ArgsRegistry.userConsentActionTypes];
+    
     return all;
 }
 
@@ -110,7 +111,13 @@
         @"addCommentToFeature": @(IBGActionAddCommentToFeature),
     };
 }
-
++ (ArgsDictionary *) userConsentActionTypes {
+    return @{
+        @"dropAutoCapturedMedia": @(IBGActionTypeDropAutoCapturedMedia),
+        @"dropLogs": @(IBGActionTypeDropLogs),
+        @"noChat": @(IBGActionTypeNoChat)
+    };
+}
 + (ArgsDictionary *) extendedBugReportStates {
     return @{
         @"enabledWithRequiredFields": @(IBGExtendedBugReportModeEnabledWithRequiredFields),